+++ commitlint.config.js
@@ -4,7 +4,7 @@ module.exports = {
     'type-enum': [
       2,
       'always',
-      ['build', 'ci', 'docs', 'feat', 'fix', 'perf', 'refactor', 'revert', 'style', 'test', 'lint', 'example', 'chore'],
+      ['build', 'ci', 'doc', 'feat', 'fix', 'perf', 'refactor', 'revert', 'style', 'test', 'lint', 'example', 'chore'],
     ],
   },
 };