+++ CHANGELOG.md
@@ -1,3 +1,46 @@
+## [1.3.2](https://github.com/pchmn/expo-material3-theme/compare/v1.3.1...v1.3.2) (2024-01-15)
+
+
+### Bug Fixes
+
+* support android gradle plugin version >= 8 ([d7cbf1e](https://github.com/pchmn/expo-material3-theme/commit/d7cbf1e51278ee77fb325dd45c02f706753fb120))
+
+## [1.3.1](https://github.com/pchmn/expo-material3-theme/compare/v1.3.0...v1.3.1) (2023-07-11)
+
+
+### Bug Fixes
+
+* typo (`surfaceContainerHight` -> `surfaceContainerHigh`) ([1a64d09](https://github.com/pchmn/expo-material3-theme/commit/1a64d094bc7d08cf0864a42ffa8ea9a96b977086))
+
+# [1.3.0](https://github.com/pchmn/expo-material3-theme/compare/v1.2.0...v1.3.0) (2023-07-11)
+
+
+### Features
+
+* export `isDynamicThemeSupported` and `Material3Scheme` ([a91a105](https://github.com/pchmn/expo-material3-theme/commit/a91a1050582f84a34d61575dc4ffb9706609e61a))
+
+# [1.2.0](https://github.com/pchmn/expo-material3-theme/compare/v1.1.0...v1.2.0) (2023-07-09)
+
+
+### Features
+
+* add new color tokens ([efcabe0](https://github.com/pchmn/expo-material3-theme/commit/efcabe05da8db8ffc690559a7f02d7360f834715))
+
+# [1.1.0](https://github.com/pchmn/expo-material3-theme/compare/v1.0.1...v1.1.0) (2023-07-07)
+
+* fix gradle 8 compatibility ([9f04784](https://github.com/pchmn/expo-material3-theme/commit/9f04784fadfdde8c80fd3c889e72f55267fee156))
+* use @material/material-color-utilities ([d2f6085](https://github.com/pchmn/expo-material3-theme/commit/d2f6085e4ed3cf207ef534d71ab4c38ec8311329))
+
+## [1.0.1](https://github.com/pchmn/expo-material3-theme/compare/v1.0.0...v1.0.1) (2023-04-03)
+
+
+### Bug Fixes
+
+* exclude unnecessary files from package ([a5d08df](https://github.com/pchmn/expo-material3-theme/commit/a5d08dfc7a9c356afe276ea5271adc6a3363ec93))
+
+# 1.0.0 (2023-03-31)
+
+
 # [1.0.0-beta.2](https://github.com/pchmn/expo-material3-theme/compare/v1.0.0-beta.1...v1.0.0-beta.2) (2023-03-30)
 
 
@@ -9,10 +52,8 @@
 
 ### BREAKING CHANGES
 
-* remove `overwrite` param
-
-`overwrite` param in `useMaterial3Theme` is removed, use `sourceColor` param to overwrite system theme
+* `overwrite` param in `useMaterial3Theme` is removed, use `sourceColor` param to overwrite system theme
 
 # 1.0.0-beta.1 (2023-03-16)
 
-First version
+First beta version